+++ src/manifest.js
@@ -6,7 +6,7 @@ import { assign } from './utils'
  * @param {Object} obj
  *   @param {String} obj.host
  *   @param {boolean} obj.allowResourceHostOverride - default: false
- *   @param {boolean} obj.disableParamsEncode - default: false
+ *   @param {Function} obj.paramsEncodingOverride - default: encodeURIComponent
  *   @param {Object} obj.gatewayConfigs - default: base values from mappersmith
  *   @param {Object} obj.ignoreGlobalMiddleware - default: false
  *   @param {Object} obj.resources - default: {}
@@ -16,7 +16,7 @@ import { assign } from './utils'
 function Manifest(obj, { gatewayConfigs = null, middleware = [], context = {} }) {
   this.host = obj.host
   this.allowResourceHostOverride = obj.allowResourceHostOverride || false
-  this.disableParamsEncode = obj.disableParamsEncode || false
+  this.paramsEncodingOverride = obj.paramsEncodingOverride || encodeURIComponent
   this.bodyAttr = obj.bodyAttr
   this.headersAttr = obj.headersAttr
   this.authAttr = obj.authAttr
@@ -67,7 +67,7 @@ Manifest.prototype = {
         {
           host: this.host,
           allowResourceHostOverride: this.allowResourceHostOverride,
-          disableParamsEncode: this.disableParamsEncode,
+          paramsEncodingOverride: this.paramsEncodingOverride,
           bodyAttr: this.bodyAttr,
           headersAttr: this.headersAttr,
           authAttr: this.authAttr,