+++ src/request.ts
@@ -102,9 +102,7 @@ export class Request {
       const pattern = new RegExp(`{${key}\\??}`, 'g')
       const value = params[key]
       if (value != null && typeof value !== 'object') {
-        const valueToReplace = this.methodDescriptor.disableParamsEncode
-          ? value.toString()
-          : encodeURIComponent(value)
+        const valueToReplace = this.methodDescriptor.paramsEncodingOverride(value)
         path = path.replace(pattern, valueToReplace)
         delete params[key]
       }